git-svn-id: https://kbengine.svn.sourceforge.net/svnroot/kbengine@697 d26bdfbe-e24b-4f15-971f-8e0596e1bc08

+++ kbe/src/lib/entitydef/datatype.hpp
@@ -284,7 +284,7 @@ public:
 	const char* getName(void)const{ return name_.c_str();}
 protected:
 	std::string name_;
-	uint8 elemCount_;
+	uint32 elemCount_;
 };
 
 class StringType : public DataType
